TFP-3040: Gjør det mulig å legge til inntekt fra andre enheter på inntektspost (eksempel ved juridisk enhet) uten å miste inntekt som hører til den enheten
