+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/IAYRegisterInnhentingFellesTjenesteImpl.java
@@ -62,15 +62,14 @@ import no.nav.vedtak.felles.integrasjon.aktør.klient.AktørConsumer;
 public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegisterInnhentingTjeneste {
 
     public static final Map<RegisterdataElement, InntektsKilde> ELEMENT_TIL_INNTEKTS_KILDE_MAP = Map.of(RegisterdataElement.INNTEKT_PENSJONSGIVENDE, InntektsKilde.INNTEKT_OPPTJENING, RegisterdataElement.INNTEKT_BEREGNINGSGRUNNLAG, InntektsKilde.INNTEKT_BEREGNING, RegisterdataElement.INNTEKT_SAMMENLIGNINGSGRUNNLAG, InntektsKilde.INNTEKT_SAMMENLIGNING);
-    private static final Logger LOGGER = LoggerFactory.getLogger(IAYRegisterInnhentingFellesTjenesteImpl.class);
     private VirksomhetTjeneste virksomhetTjeneste;
     private YtelseRegisterInnhenting ytelseRegisterInnhenting;
     private InntektArbeidYtelseTjeneste inntektArbeidYtelseTjeneste;
     private InnhentingSamletTjeneste innhentingSamletTjeneste;
     private KodeverkRepository kodeverkRepository;
     private ByggYrkesaktiviteterTjeneste byggYrkesaktiviteterTjeneste;
-    private AktørConsumer aktørConsumer;
     private SigrunTjeneste sigrunTjeneste;
+    private InntektMapper inntektMapper;
 
     protected IAYRegisterInnhentingFellesTjenesteImpl() {
     }
@@ -79,36 +78,17 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
                                                       KodeverkRepository kodeverkRepository,
                                                       VirksomhetTjeneste virksomhetTjeneste,
                                                       InnhentingSamletTjeneste innhentingSamletTjeneste,
-                                                      AktørConsumer aktørConsumer, SigrunTjeneste sigrunTjeneste, VedtakYtelseRepository vedtakYtelseRepository) {
+                                                      SigrunTjeneste sigrunTjeneste,
+                                                      VedtakYtelseRepository vedtakYtelseRepository,
+                                                      InntektMapper inntektMapper) {
         this.inntektArbeidYtelseTjeneste = inntektArbeidYtelseTjeneste;
         this.kodeverkRepository = kodeverkRepository;
         this.virksomhetTjeneste = virksomhetTjeneste;
         this.innhentingSamletTjeneste = innhentingSamletTjeneste;
-        this.aktørConsumer = aktørConsumer;
         this.sigrunTjeneste = sigrunTjeneste;
         this.ytelseRegisterInnhenting = new YtelseRegisterInnhenting(innhentingSamletTjeneste, vedtakYtelseRepository);
         this.byggYrkesaktiviteterTjeneste = new ByggYrkesaktiviteterTjeneste(kodeverkRepository);
-    }
-
-    private void innhentNæringsOpplysninger(Kobling kobling, InntektArbeidYtelseAggregatBuilder inntektArbeidYtelseAggregatBuilder) {
-        Map<IntervallEntitet, Map<InntektspostType, BigDecimal>> map = sigrunTjeneste.beregnetSkatt(kobling.getAktørId());
-        InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder = inntektArbeidYtelseAggregatBuilder
-            .getAktørInntektBuilder(kobling.getAktørId());
-
-        InntektBuilder inntektBuilder = aktørInntektBuilder.getInntektBuilder(InntektsKilde.SIGRUN, null);
-
-        for (Map.Entry<IntervallEntitet, Map<InntektspostType, BigDecimal>> entry : map.entrySet()) {
-            for (Map.Entry<InntektspostType, BigDecimal> type : entry.getValue().entrySet()) {
-                InntektspostBuilder inntektspostBuilder = inntektBuilder.getInntektspostBuilder();
-                inntektspostBuilder
-                    .medInntektspostType(type.getKey())
-                    .medBeløp(type.getValue())
-                    .medPeriode(entry.getKey().getFomDato(), entry.getKey().getTomDato());
-                inntektBuilder.leggTilInntektspost(inntektspostBuilder);
-            }
-        }
-        aktørInntektBuilder.leggTilInntekt(inntektBuilder);
-        inntektArbeidYtelseAggregatBuilder.leggTilAktørInntekt(aktørInntektBuilder);
+        this.inntektMapper = inntektMapper;
     }
 
     @Override
@@ -145,12 +125,9 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
         return grunnlagBuilder;
     }
 
-    private List<ArbeidsforholdOverstyring> finnOverstyringerForBortfalteArbeidsforhold(Set<ArbeidsforholdIdentifikator> innhentetArbeidsforhold, ArbeidsforholdInformasjon informasjon) {
-        return informasjon.getOverstyringer().stream()
-            .filter(ov -> innhentetArbeidsforhold.stream()
-                .noneMatch(arbeid -> arbeid.getArbeidsgiver().getIdentifikator().equals(ov.getArbeidsgiver().getIdentifikator())
-                    && (arbeid.harArbeidsforholdRef() && arbeid.getArbeidsforholdId().gjelderFor(informasjon.finnEkstern(ov.getArbeidsgiver(), ov.getArbeidsforholdRef())))))
-            .collect(Collectors.toList());
+    private void innhentNæringsOpplysninger(Kobling kobling, InntektArbeidYtelseAggregatBuilder inntektArbeidYtelseAggregatBuilder) {
+        Map<IntervallEntitet, Map<InntektspostType, BigDecimal>> map = sigrunTjeneste.beregnetSkatt(kobling.getAktørId());
+        inntektMapper.mapFraSigrun(kobling.getAktørId(), map, inntektArbeidYtelseAggregatBuilder);
     }
 
     private void innhentYtelser(Kobling kobling, InntektArbeidYtelseAggregatBuilder builder) {
@@ -162,31 +139,6 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
 
     private Set<ArbeidsforholdIdentifikator> innhentArbeidsforhold(Kobling kobling, InntektArbeidYtelseAggregatBuilder builder, Set<RegisterdataElement> informasjonsElementer) {
         return byggOpptjeningOpplysningene(kobling, kobling.getAktørId(), kobling.getOpplysningsperiode().tilIntervall(), builder, informasjonsElementer);
-    }
-
-    private void leggTilInntekter(AktørId aktørId, InntektArbeidYtelseAggregatBuilder builder, InntektsInformasjon inntektsInformasjon) {
-        InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder = builder.getAktørInntektBuilder(aktørId);
-        InntektsKilde kilde = inntektsInformasjon.getKilde();
-        aktørInntektBuilder.fjernInntekterFraKilde(kilde);
-
-        inntektsInformasjon.getMånedsinntekterGruppertPåArbeidsgiver()
-            .forEach((identifikator, inntektOgRegelListe) -> leggTilInntekterPåArbeidsforhold(builder, aktørInntektBuilder, inntektOgRegelListe, identifikator,
-                kilde));
-
-        final List<Månedsinntekt> ytelsesTrygdEllerPensjonInntekt = inntektsInformasjon.getYtelsesTrygdEllerPensjonInntektSummert();
-        if (!ytelsesTrygdEllerPensjonInntekt.isEmpty()) {
-            leggTilYtelseInntekter(ytelsesTrygdEllerPensjonInntekt, builder, aktørId, kilde);
-        }
-    }
-
-    private void leggTilYtelseInntekter(List<Månedsinntekt> ytelsesTrygdEllerPensjonInntekt, InntektArbeidYtelseAggregatBuilder builder, AktørId aktørId,
-                                        InntektsKilde inntektOpptjening) {
-        final InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder = builder.getAktørInntektBuilder(aktørId);
-        final InntektBuilder inntektBuilderForYtelser = aktørInntektBuilder.getInntektBuilderForYtelser(inntektOpptjening);
-        ytelsesTrygdEllerPensjonInntekt.forEach(mi -> lagInntektsposterYtelse(mi, inntektBuilderForYtelser));
-
-        aktørInntektBuilder.leggTilInntekt(inntektBuilderForYtelser);
-        builder.leggTilAktørInntekt(aktørInntektBuilder);
     }
 
     private void oversettFrilanseArbeidsforhold(Kobling kobling,
@@ -220,44 +172,6 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
 
         aktørArbeidBuilder.leggTilYrkesaktivitet(yrkesaktivitetBuilder);
         builder.leggTilAktørArbeid(aktørArbeidBuilder);
-    }
-
-    private void leggTilInntekterPåArbeidsforhold(InntektArbeidYtelseAggregatBuilder builder,
-                                                  InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
-                                                  Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> månedsinntekterGruppertPåArbeidsgiver,
-                                                  String arbeidsgiverIdentifikator, InntektsKilde inntektOpptjening) {
-
-        Arbeidsgiver arbeidsgiver;
-        if (OrganisasjonsNummerValidator.erGyldig(arbeidsgiverIdentifikator)) {
-            boolean orgledd = virksomhetTjeneste.sjekkOmVirksomhetErOrgledd(arbeidsgiverIdentifikator);
-            if (!orgledd) {
-                LocalDate hentedato = finnHentedatoForJuridisk(månedsinntekterGruppertPåArbeidsgiver.keySet());
-                arbeidsgiver = Arbeidsgiver.virksomhet(virksomhetTjeneste.hentOgLagreOrganisasjonMedHensynTilJuridisk(arbeidsgiverIdentifikator, hentedato));
-                aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening));
-                builder.leggTilAktørInntekt(aktørInntektBuilder);
-            } else {
-                LOGGER.info("Inntekter rapportert på orglegg({}), blir IKKE lagret", arbeidsgiverIdentifikator);
-            }
-        } else {
-            if (PersonIdent.erGyldigFnr(arbeidsgiverIdentifikator)) {
-                Optional<String> arbeidsgiverOpt = aktørConsumer.hentAktørIdForPersonIdent(arbeidsgiverIdentifikator);
-                if (arbeidsgiverOpt.isEmpty()) {
-                    throw InnhentingFeil.FACTORY.finnerIkkeAktørIdForArbeidsgiverSomErPrivatperson().toException();
-                }
-                arbeidsgiver = Arbeidsgiver.person(new AktørId(arbeidsgiverOpt.get()));
-            } else {
-                LOGGER.info("Arbeidsgiveridentifikator: {}", arbeidsgiverIdentifikator);
-                arbeidsgiver = Arbeidsgiver.person(new AktørId(arbeidsgiverIdentifikator));
-            }
-            aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening));
-            builder.leggTilAktørInntekt(aktørInntektBuilder);
-        }
-    }
-
-    private LocalDate finnHentedatoForJuridisk(Set<YearMonth> inntekterForMåneder) {
-        return inntekterForMåneder.stream()
-            .map(m -> LocalDate.of(m.getYear(), m.getMonth(), 1))
-            .max(Comparator.naturalOrder()).orElse(LocalDate.now());
     }
 
     private Set<ArbeidsforholdIdentifikator> byggOpptjeningOpplysningene(Kobling kobling, AktørId aktørId, Interval opplysningsPeriode,
@@ -298,7 +212,7 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
             kobling.getOpplysningsperiode().tilIntervall(), inntektsKilde);
 
         if (informasjonsElementer.contains(registerdataElement)) {
-            leggTilInntekter(aktørId, builder, inntektsInformasjon);
+            inntektMapper.mapFraInntektskomponent(aktørId, builder, inntektsInformasjon);
         }
 
         if (inntektsKilde.equals(InntektsKilde.INNTEKT_OPPTJENING) && informasjonsElementer.contains(RegisterdataElement.ARBEIDSFORHOLD)) {
@@ -361,78 +275,5 @@ public abstract class IAYRegisterInnhentingFellesTjenesteImpl implements IAYRegi
             periode = IntervallEntitet.fraOgMedTilOgMed(frilansArbeidsforhold.getFom(), frilansArbeidsforhold.getTom());
         }
         return yrkesaktivitetBuilder.getAktivitetsAvtaleBuilder(periode, true);
-    }
-
-    private InntektBuilder byggInntekt(Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> inntekter,
-                                       Arbeidsgiver arbeidsgiver,
-                                       InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
-                                       InntektsKilde inntektOpptjening) {
-
-        InntektBuilder inntektBuilder = aktørInntektBuilder.getInntektBuilder(inntektOpptjening, new Opptjeningsnøkkel(arbeidsgiver));
-
-        for (YearMonth måned : inntekter.keySet()) {
-            List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel> månedsinnteker = inntekter.get(måned);
-            Map<String, Integer> antalInntekterForAvgiftsregel = månedsinnteker
-                .stream()
-                .filter(e -> e.getSkatteOgAvgiftsregelType() != null)
-                .collect(Collectors.groupingBy(
-                    InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getSkatteOgAvgiftsregelType,
-                    Collectors.collectingAndThen(
-                        Collectors.mapping(InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getBeløp, Collectors.toSet()),
-                        Set::size)));
-
-            Optional<String> valgtSkatteOgAvgiftsregel = Optional.empty();
-            BigDecimal beløpSum = månedsinnteker.stream().map(InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getBeløp).reduce(BigDecimal.ZERO,
-                BigDecimal::add);
-            if (antalInntekterForAvgiftsregel.keySet().size() > 1) {
-                String skatteOgAvgiftsregler = antalInntekterForAvgiftsregel.keySet().stream().collect(Collectors.joining(", "));
-                // TODO Diamant velger her en random verdi.
-                valgtSkatteOgAvgiftsregel = Optional.of(antalInntekterForAvgiftsregel.keySet().iterator().next());
-                LOGGER.error("ArbeidsgiverEntitet orgnr {} har flere månedsinntekter for måned {} med forskjellige skatte -og avgiftsregler {}. Velger {}",
-                    arbeidsgiver.getIdentifikator(), måned, skatteOgAvgiftsregler, valgtSkatteOgAvgiftsregel);
-            } else if (antalInntekterForAvgiftsregel.keySet().size() == 1) {
-                valgtSkatteOgAvgiftsregel = Optional.of(antalInntekterForAvgiftsregel.keySet().iterator().next());
-            }
-
-            lagInntektsposter(måned, beløpSum, valgtSkatteOgAvgiftsregel, inntektBuilder);
-        }
-
-        return inntektBuilder
-            .medArbeidsgiver(arbeidsgiver);
-    }
-
-    private void lagInntektsposterYtelse(Månedsinntekt månedsinntekt, InntektBuilder inntektBuilder) {
-        inntektBuilder.leggTilInntektspost(inntektBuilder.getInntektspostBuilder()
-            .medBeløp(månedsinntekt.getBeløp())
-            .medPeriode(månedsinntekt.getMåned().atDay(1), månedsinntekt.getMåned().atEndOfMonth())
-            .medInntektspostType(InntektspostType.YTELSE)
-            .medYtelse(mapTilKodeliste(månedsinntekt)));
-    }
-
-    private void lagInntektsposter(YearMonth måned, BigDecimal sumInntektsbeløp, Optional<String> valgtSkatteOgAvgiftsregel,
-                                   InntektBuilder inntektBuilder) {
-
-        InntektspostBuilder inntektspostBuilder = inntektBuilder.getInntektspostBuilder();
-        inntektspostBuilder
-            .medBeløp(sumInntektsbeløp)
-            .medPeriode(måned.atDay(1), måned.atEndOfMonth())
-            .medInntektspostType(InntektspostType.LØNN);
-
-        if (valgtSkatteOgAvgiftsregel.isPresent()) {
-            SkatteOgAvgiftsregelType skatteOgAvgiftsregelType = kodeverkRepository.finnForKodeverkEiersKode(SkatteOgAvgiftsregelType.class,
-                valgtSkatteOgAvgiftsregel.get());
-            inntektspostBuilder.medSkatteOgAvgiftsregelType(skatteOgAvgiftsregelType);
-        }
-
-        inntektBuilder.leggTilInntektspost(inntektspostBuilder);
-    }
-
-    private YtelseInntektspostType mapTilKodeliste(Månedsinntekt månedsinntekt) {
-        if (månedsinntekt.getPensjonKode() != null) {
-            return kodeverkRepository.finnForKodeverkEiersKode(PensjonTrygdType.class, månedsinntekt.getPensjonKode());
-        } else if (månedsinntekt.getYtelseKode() != null) {
-            return kodeverkRepository.finnForKodeverkEiersKode(OffentligYtelseType.class, månedsinntekt.getYtelseKode());
-        }
-        return kodeverkRepository.finnForKodeverkEiersKode(NæringsinntektType.class, månedsinntekt.getNæringsinntektKode());
     }
 }
