+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/InntektMapper.java
@@ -1,0 +1,228 @@
+package no.nav.foreldrepenger.abakus.registerdata;
+
+import no.nav.foreldrepenger.abakus.domene.iay.Arbeidsgiver;
+import no.nav.foreldrepenger.abakus.domene.iay.InntektArbeidYtelseAggregatBuilder;
+import no.nav.foreldrepenger.abakus.domene.iay.InntektBuilder;
+import no.nav.foreldrepenger.abakus.domene.iay.InntektspostBuilder;
+import no.nav.foreldrepenger.abakus.domene.iay.NæringsinntektType;
+import no.nav.foreldrepenger.abakus.domene.iay.OffentligYtelseType;
+import no.nav.foreldrepenger.abakus.domene.iay.Opptjeningsnøkkel;
+import no.nav.foreldrepenger.abakus.domene.iay.PensjonTrygdType;
+import no.nav.foreldrepenger.abakus.domene.iay.YtelseInntektspostType;
+import no.nav.foreldrepenger.abakus.domene.iay.kodeverk.InntektsKilde;
+import no.nav.foreldrepenger.abakus.domene.iay.kodeverk.InntektspostType;
+import no.nav.foreldrepenger.abakus.domene.iay.kodeverk.SkatteOgAvgiftsregelType;
+import no.nav.foreldrepenger.abakus.felles.jpa.IntervallEntitet;
+import no.nav.foreldrepenger.abakus.kodeverk.KodeverkRepository;
+import no.nav.foreldrepenger.abakus.registerdata.arbeidsgiver.virksomhet.VirksomhetTjeneste;
+import no.nav.foreldrepenger.abakus.registerdata.inntekt.komponenten.InntektsInformasjon;
+import no.nav.foreldrepenger.abakus.registerdata.inntekt.komponenten.Månedsinntekt;
+import no.nav.foreldrepenger.abakus.typer.AktørId;
+import no.nav.foreldrepenger.abakus.typer.OrganisasjonsNummerValidator;
+import no.nav.foreldrepenger.abakus.typer.PersonIdent;
+import no.nav.vedtak.felles.integrasjon.aktør.klient.AktørConsumer;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
+import java.math.BigDecimal;
+import java.time.LocalDate;
+import java.time.YearMonth;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+@ApplicationScoped
+public class InntektMapper {
+    private static final Logger LOGGER = LoggerFactory.getLogger(InntektMapper.class);
+    private VirksomhetTjeneste virksomhetTjeneste;
+    private KodeverkRepository kodeverkRepository;
+    private AktørConsumer aktørConsumer;
+
+    public InntektMapper() {
+        // For CDI
+    }
+
+    @Inject
+    protected InntektMapper(KodeverkRepository kodeverkRepository,
+                            VirksomhetTjeneste virksomhetTjeneste,
+                            AktørConsumer aktørConsumer) {
+        this.kodeverkRepository = kodeverkRepository;
+        this.virksomhetTjeneste = virksomhetTjeneste;
+        this.aktørConsumer = aktørConsumer;
+    }
+
+    public void mapFraSigrun(AktørId aktørId, Map<IntervallEntitet, Map<InntektspostType, BigDecimal>> map, InntektArbeidYtelseAggregatBuilder inntektArbeidYtelseAggregatBuilder) {
+        InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder = inntektArbeidYtelseAggregatBuilder
+            .getAktørInntektBuilder(aktørId);
+
+        InntektBuilder inntektBuilder = aktørInntektBuilder.getInntektBuilder(InntektsKilde.SIGRUN, null);
+
+        for (Map.Entry<IntervallEntitet, Map<InntektspostType, BigDecimal>> entry : map.entrySet()) {
+            for (Map.Entry<InntektspostType, BigDecimal> type : entry.getValue().entrySet()) {
+                InntektspostBuilder inntektspostBuilder = inntektBuilder.getInntektspostBuilder();
+                inntektspostBuilder
+                    .medInntektspostType(type.getKey())
+                    .medBeløp(type.getValue())
+                    .medPeriode(entry.getKey().getFomDato(), entry.getKey().getTomDato());
+                inntektBuilder.leggTilInntektspost(inntektspostBuilder);
+            }
+        }
+        aktørInntektBuilder.leggTilInntekt(inntektBuilder);
+        inntektArbeidYtelseAggregatBuilder.leggTilAktørInntekt(aktørInntektBuilder);
+    }
+
+    public void mapFraInntektskomponent(AktørId aktørId, InntektArbeidYtelseAggregatBuilder builder, InntektsInformasjon inntektsInformasjon) {
+        InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder = builder.getAktørInntektBuilder(aktørId);
+        InntektsKilde kilde = inntektsInformasjon.getKilde();
+        aktørInntektBuilder.fjernInntekterFraKilde(kilde);
+
+        inntektsInformasjon.getMånedsinntekterGruppertPåArbeidsgiver()
+            .forEach((identifikator, inntektOgRegelListe) -> leggTilInntekterPåArbeidsforhold(builder, aktørInntektBuilder, inntektOgRegelListe, identifikator,
+                kilde)); // Loopes 3 ganger.
+
+        final List<Månedsinntekt> ytelsesTrygdEllerPensjonInntekt = inntektsInformasjon.getYtelsesTrygdEllerPensjonInntektSummert();
+        if (!ytelsesTrygdEllerPensjonInntekt.isEmpty()) {
+            leggTilYtelseInntekter(ytelsesTrygdEllerPensjonInntekt, builder, aktørId, kilde);
+        }
+    }
+
+    private void leggTilYtelseInntekter(List<Månedsinntekt> ytelsesTrygdEllerPensjonInntekt, InntektArbeidYtelseAggregatBuilder builder, AktørId aktørId,
+                                        InntektsKilde inntektOpptjening) {
+        final InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder = builder.getAktørInntektBuilder(aktørId);
+        final InntektBuilder inntektBuilderForYtelser = aktørInntektBuilder.getInntektBuilderForYtelser(inntektOpptjening);
+        ytelsesTrygdEllerPensjonInntekt.forEach(mi -> lagInntektsposterYtelse(mi, inntektBuilderForYtelser));
+
+        aktørInntektBuilder.leggTilInntekt(inntektBuilderForYtelser);
+        builder.leggTilAktørInntekt(aktørInntektBuilder);
+    }
+
+    private void leggTilInntekterPåArbeidsforhold(InntektArbeidYtelseAggregatBuilder builder,
+                                                  InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
+                                                  Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> månedsinntekterGruppertPåArbeidsgiver,
+                                                  String arbeidsgiverIdentifikator, InntektsKilde inntektOpptjening) {
+
+        Arbeidsgiver arbeidsgiver;
+        if (OrganisasjonsNummerValidator.erGyldig(arbeidsgiverIdentifikator)) {
+            boolean orgledd = virksomhetTjeneste.sjekkOmVirksomhetErOrgledd(arbeidsgiverIdentifikator);
+            if (!orgledd) {
+                LocalDate hentedato = finnHentedatoForJuridisk(månedsinntekterGruppertPåArbeidsgiver.keySet());
+                arbeidsgiver = Arbeidsgiver.virksomhet(virksomhetTjeneste.hentOgLagreOrganisasjonMedHensynTilJuridisk(arbeidsgiverIdentifikator, hentedato)); // Orgnr som har 650 inntekt vil returnere et annet orgnr her
+                if (!arbeidsgiver.getOrgnr().equals(arbeidsgiverIdentifikator)) {
+                    aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening, arbeidsgiverIdentifikator));
+                } else {
+                    aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening));
+                }
+                builder.leggTilAktørInntekt(aktørInntektBuilder);
+            } else {
+                LOGGER.info("Inntekter rapportert på orglegg({}), blir IKKE lagret", arbeidsgiverIdentifikator);
+            }
+        } else {
+            if (PersonIdent.erGyldigFnr(arbeidsgiverIdentifikator)) {
+                Optional<String> arbeidsgiverOpt = aktørConsumer.hentAktørIdForPersonIdent(arbeidsgiverIdentifikator);
+                if (arbeidsgiverOpt.isEmpty()) {
+                    throw InnhentingFeil.FACTORY.finnerIkkeAktørIdForArbeidsgiverSomErPrivatperson().toException();
+                }
+                arbeidsgiver = Arbeidsgiver.person(new AktørId(arbeidsgiverOpt.get()));
+            } else {
+                LOGGER.info("Arbeidsgiveridentifikator: {}", arbeidsgiverIdentifikator);
+                arbeidsgiver = Arbeidsgiver.person(new AktørId(arbeidsgiverIdentifikator));
+            }
+            aktørInntektBuilder.leggTilInntekt(byggInntekt(månedsinntekterGruppertPåArbeidsgiver, arbeidsgiver, aktørInntektBuilder, inntektOpptjening));
+            builder.leggTilAktørInntekt(aktørInntektBuilder);
+        }
+    }
+
+    private LocalDate finnHentedatoForJuridisk(Set<YearMonth> inntekterForMåneder) {
+        return inntekterForMåneder.stream()
+            .map(m -> LocalDate.of(m.getYear(), m.getMonth(), 1))
+            .max(Comparator.naturalOrder()).orElse(LocalDate.now());
+    }
+
+    private InntektBuilder byggInntekt(Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> inntekter,
+                                       Arbeidsgiver arbeidsgiver,
+                                       InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
+                                       InntektsKilde inntektOpptjening) {
+        return byggInntekt(inntekter, arbeidsgiver, aktørInntektBuilder, inntektOpptjening, null);
+    }
+
+
+    private InntektBuilder byggInntekt(Map<YearMonth, List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel>> inntekter,
+                                       Arbeidsgiver arbeidsgiver,
+                                       InntektArbeidYtelseAggregatBuilder.AktørInntektBuilder aktørInntektBuilder,
+                                       InntektsKilde inntektOpptjening, String opprineligUtbetalerId) {
+        InntektBuilder inntektBuilder = aktørInntektBuilder.getInntektBuilder(inntektOpptjening, new Opptjeningsnøkkel(arbeidsgiver));
+
+        for (YearMonth måned : inntekter.keySet()) {
+            List<InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel> månedsinnteker = inntekter.get(måned);
+            Map<String, Integer> antalInntekterForAvgiftsregel = månedsinnteker
+                .stream()
+                .filter(e -> e.getSkatteOgAvgiftsregelType() != null)
+                .collect(Collectors.groupingBy(
+                    InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getSkatteOgAvgiftsregelType,
+                    Collectors.collectingAndThen(
+                        Collectors.mapping(InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getBeløp, Collectors.toSet()),
+                        Set::size)));
+
+            Optional<String> valgtSkatteOgAvgiftsregel = Optional.empty();
+            BigDecimal beløpSum = månedsinnteker.stream().map(InntektsInformasjon.MånedsbeløpOgSkatteOgAvgiftsregel::getBeløp).reduce(BigDecimal.ZERO,
+                BigDecimal::add);
+            if (antalInntekterForAvgiftsregel.keySet().size() > 1) {
+                String skatteOgAvgiftsregler = antalInntekterForAvgiftsregel.keySet().stream().collect(Collectors.joining(", "));
+                // TODO Diamant velger her en random verdi.
+                valgtSkatteOgAvgiftsregel = Optional.of(antalInntekterForAvgiftsregel.keySet().iterator().next());
+                LOGGER.error("ArbeidsgiverEntitet orgnr {} har flere månedsinntekter for måned {} med forskjellige skatte -og avgiftsregler {}. Velger {}",
+                    arbeidsgiver.getIdentifikator(), måned, skatteOgAvgiftsregler, valgtSkatteOgAvgiftsregel);
+            } else if (antalInntekterForAvgiftsregel.keySet().size() == 1) {
+                valgtSkatteOgAvgiftsregel = Optional.of(antalInntekterForAvgiftsregel.keySet().iterator().next());
+            }
+
+            lagInntektsposter(måned, beløpSum, valgtSkatteOgAvgiftsregel, inntektBuilder, opprineligUtbetalerId);
+        }
+
+        return inntektBuilder
+            .medArbeidsgiver(arbeidsgiver);
+    }
+
+    private void lagInntektsposterYtelse(Månedsinntekt månedsinntekt, InntektBuilder inntektBuilder) {
+        inntektBuilder.leggTilInntektspost(inntektBuilder.getInntektspostBuilder()
+            .medBeløp(månedsinntekt.getBeløp())
+            .medPeriode(månedsinntekt.getMåned().atDay(1), månedsinntekt.getMåned().atEndOfMonth())
+            .medInntektspostType(InntektspostType.YTELSE)
+            .medYtelse(mapTilKodeliste(månedsinntekt)));
+    }
+
+    private void lagInntektsposter(YearMonth måned, BigDecimal sumInntektsbeløp, Optional<String> valgtSkatteOgAvgiftsregel,
+                               InntektBuilder inntektBuilder, String originalUtbetalerId) {
+
+        InntektspostBuilder inntektspostBuilder = inntektBuilder.getInntektspostBuilder();
+        inntektspostBuilder
+            .medBeløp(sumInntektsbeløp)
+            .medPeriode(måned.atDay(1), måned.atEndOfMonth())
+            .medInntektspostType(InntektspostType.LØNN);
+        if (originalUtbetalerId != null) {
+            inntektspostBuilder.medOriginalUtbetalerId(originalUtbetalerId);
+        }
+
+        if (valgtSkatteOgAvgiftsregel.isPresent()) {
+            SkatteOgAvgiftsregelType skatteOgAvgiftsregelType = kodeverkRepository.finnForKodeverkEiersKode(SkatteOgAvgiftsregelType.class,
+                valgtSkatteOgAvgiftsregel.get());
+            inntektspostBuilder.medSkatteOgAvgiftsregelType(skatteOgAvgiftsregelType);
+        }
+
+        inntektBuilder.leggTilInntektspost(inntektspostBuilder);
+    }
+
+    private YtelseInntektspostType mapTilKodeliste(Månedsinntekt månedsinntekt) {
+        if (månedsinntekt.getPensjonKode() != null) {
+            return kodeverkRepository.finnForKodeverkEiersKode(PensjonTrygdType.class, månedsinntekt.getPensjonKode());
+        } else if (månedsinntekt.getYtelseKode() != null) {
+            return kodeverkRepository.finnForKodeverkEiersKode(OffentligYtelseType.class, månedsinntekt.getYtelseKode());
+        }
+        return kodeverkRepository.finnForKodeverkEiersKode(NæringsinntektType.class, månedsinntekt.getNæringsinntektKode());
+    }
+}
