+++ domenetjenester/iay/src/main/java/no/nav/foreldrepenger/abakus/registerdata/es/IAYRegisterInnhentingESTjenesteImpl.java
@@ -9,6 +9,7 @@ import no.nav.foreldrepenger.abakus.kobling.kontroll.YtelseTypeRef;
 import no.nav.foreldrepenger.abakus.kodeverk.KodeverkRepository;
 import no.nav.foreldrepenger.abakus.registerdata.IAYRegisterInnhentingFellesTjenesteImpl;
 import no.nav.foreldrepenger.abakus.registerdata.InnhentingSamletTjeneste;
+import no.nav.foreldrepenger.abakus.registerdata.InntektMapper;
 import no.nav.foreldrepenger.abakus.registerdata.arbeidsgiver.virksomhet.VirksomhetTjeneste;
 import no.nav.foreldrepenger.abakus.registerdata.inntekt.sigrun.SigrunTjeneste;
 import no.nav.foreldrepenger.abakus.vedtak.domene.VedtakYtelseRepository;
@@ -27,12 +28,15 @@ public class IAYRegisterInnhentingESTjenesteImpl extends IAYRegisterInnhentingFe
                                                KodeverkRepository kodeverkRepository,
                                                VirksomhetTjeneste virksomhetTjeneste,
                                                InnhentingSamletTjeneste innhentingSamletTjeneste,
-                                               AktørConsumer aktørConsumer, SigrunTjeneste sigrunTjeneste, VedtakYtelseRepository vedtakYtelseRepository) {
+                                               AktørConsumer aktørConsumer,
+                                               SigrunTjeneste sigrunTjeneste,
+                                               VedtakYtelseRepository vedtakYtelseRepository,
+                                               InntektMapper inntektMapper) {
         super(inntektArbeidYtelseTjeneste,
             kodeverkRepository,
             virksomhetTjeneste,
             innhentingSamletTjeneste,
-            aktørConsumer, sigrunTjeneste, vedtakYtelseRepository);
+            sigrunTjeneste, vedtakYtelseRepository, inntektMapper);
     }
 
     @Override
