+++ migreringer/src/main/resources/db/migration/defaultDS/1.0/V1.1_62__oprinnelig_enhet_inntektspost.sql
@@ -1,0 +1,2 @@
+ALTER TABLE iay_inntektspost
+    ADD COLUMN opprinnelig_utbetaler_id VARCHAR(19);
